Resset password seperate component
