+++ gui/src/components/preboarding/login.vue
@@ -86,110 +86,15 @@
         </form>
       </div>
     </v-container>
-    
-    <v-dialog
-      v-model="showResetPasswordDialog"
-      persistent
-      max-width="500"
-      id="reset-password-form"
-      >
-      <v-card>
-        <v-card-title class="title mt-6 ml-3">
-          <span>{{ $t("login.change-password") }}</span>
-        </v-card-title>
-        <v-divider />
-        <v-col class="col-6 ml-7 pb-0">
-          <div
-            class="cortx-form-group"
-            :class="{
-              'cortx-form-group--error': $v.resetAccountForm.password.$error
-            }"
-          >
-            <label
-              class="cortx-form-group-label"
-              for="user-password"
-              id="new-password-label"
-            >
-              <cortx-info-tooltip
-                :label="$t('common.new-password-label')"
-                :message="passwordTooltipMessage"
-              />
-            </label>
-            <input
-              class="cortx-form__input_text"
-              type="password"
-              id="user-password"
-              name="user-password"
-              v-model.trim="resetAccountForm.password"
-              @input="$v.resetAccountForm.password.$touch"
-            />
-            <div class="cortx-form-group-label cortx-form-group-error-msg">
-              <label
-                id="new-password-required-error"
-                v-if="
-                  $v.resetAccountForm.password.$dirty &&
-                  !$v.resetAccountForm.password.required
-                "
-                >{{ $t("common.password-required") }}</label
-              >
-              <label
-                id="new-password-invalid-error"
-                v-else-if="
-                  $v.resetAccountForm.password.$dirty &&
-                  !$v.resetAccountForm.password.passwordRegex
-                "
-                >{{ $t("common.invalid-password") }}</label
-              >
-            </div>
-          </div>
-        </v-col>
-        <v-col class="col-6 ml-7 pt-0">
-          <div
-            class="cortx-form-group"
-            :class="{
-              'cortx-form-group--error':
-                $v.resetAccountForm.confirmPassword.$error
-            }"
-          >
-            <label
-              class="cortx-form-group-label"
-              for="confirm-password"
-              id="confirm-password-label"
-              >{{ $t("common.confirm-password-label") }}</label
-            >
-            <input
-              class="cortx-form__input_text"
-              type="password"
-              id="confirm-password"
-              name="confirm-password"
-              v-model.trim="resetAccountForm.confirmPassword"
-              @input="$v.resetAccountForm.confirmPassword.$touch"
-            />
-            <span
-              id="confirm-password-notmatch-error"
-              class="cortx-form-group-label cortx-form-group-error-msg"
-              v-if="
-                $v.resetAccountForm.confirmPassword.$dirty &&
-                !$v.resetAccountForm.confirmPassword.sameAsPassword
-              "
-              >{{ $t("common.password-not-match") }}</span
-            >
-          </div>
-        </v-col>
-        <v-col class="col-6 ml-7 pb-6 pt-0">
-          <button
-            type="button"
-            id="reset-password-button"
-            class="cortx-btn-primary"
-            @click="resetPassword()"
-            :disabled="$v.resetAccountForm.$invalid"
-          >
-            {{ $t("login.reset-btn") }}
-          </button>
-        </v-col>
-      </v-card>
-    </v-dialog>
-      
+
+    <cortx-reset-password
+      v-if="showResetPasswordDialog"
+      :showResetPasswordDialog="showResetPasswordDialog"
+      :authToken="authToken"
+      :username="loginForm.username"
+      @closeResetPass="closeResetPassword"
+    ></cortx-reset-password>
+
     <cortx-confirmation-dialog
       id="success-dialog"
       :show="showSuccessDialog"
@@ -204,15 +109,8 @@
 <script lang="ts">
 import { Component, Vue, Watch } from "vue-property-decorator";
 import { Validations } from "vuelidate-property-decorators";
-import { required, sameAs } from "vuelidate/lib/validators";
+import { required } from "vuelidate/lib/validators";
 import { UserLoginQueryParam } from "./../../models/user-login";
-import {
-  accountNameRegex,
-  passwordRegex,
-  passwordTooltipMessage
-} from "./../../common/regex-helpers";
-import { Api } from "../../services/api";
-import apiRegister from "../../services/api-register";
 import i18n from "./preboarding.json";
 
 @Component({
@@ -226,15 +124,10 @@ export default class CortxLogin extends Vue {
     username: "",
     password: ""
   };  
-  public resetAccountForm = {
-    password: "",
-    confirmPassword: ""
-  };
   public showResetPasswordDialog: boolean = false;
-  private passwordTooltipMessage = passwordTooltipMessage;
+  public ifResetPassword: boolean = true;
   private showSuccessDialog: boolean = false;
   private successMessage: string = "";
-  public ifResetPassword: boolean = true;
   public authToken: any;
 
   @Validations()
@@ -242,12 +135,6 @@ export default class CortxLogin extends Vue {
     loginForm: {
       username: { required },
       password: { required }
-    },
-    resetAccountForm: {
-      password: { required, passwordRegex },
-      confirmPassword: {
-        sameAsPassword: sameAs("password")
-      }
     }
   };
 
@@ -339,43 +226,12 @@ export default class CortxLogin extends Vue {
     }
   }
 
-  private async resetPassword() {
-    const updateDetails = {
-      confirmPassword: this.resetAccountForm.confirmPassword,
-      password: this.resetAccountForm.password,
-      reset_password: true
-    };
-    this.$store.dispatch(
-      "systemConfig/showLoader",
-      this.$t("login.loading-reset")
-    );
-    const res: any = await Api.patch(
-      apiRegister.csm_user,
-      updateDetails,
-      this.loginForm.username,
-      {
-        headers: {
-          auth_token: this.authToken
-        }
-      }
-    );
-    this.closeResetPasswordForm();
+  public async closeResetPassword() {
+    this.showResetPasswordDialog = false;
     this.successMessage = `${this.$t("login.password-reset-message")} ${
       this.loginForm.username
     }`;
     this.showSuccessDialog = true;
-    this.$store.dispatch("systemConfig/hideLoader");
-  }
-
-  public closeResetPasswordForm() {
-    this.resetAccountForm = {
-      password: "",
-      confirmPassword: ""
-    };
-    if (this.$v.resetAccountForm) {
-      this.$v.resetAccountForm.$reset();
-    }
-    this.showResetPasswordDialog = false;
   }
 
   public async closeSuccessDialog() {
